Move dotnet services to subdir
