+++ src/Cli/Services/Dotnet/DotnetService.cs
@@ -3,7 +3,7 @@ using Cli.Internal;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
-namespace Cli.Services
+namespace Cli.Services.Dotnet
 {
     internal abstract class DotnetService : ProcessService
     {
