+++ src/Cli/Services/Dotnet/DotnetToolService.cs
@@ -3,7 +3,7 @@ using Cli.Internal;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
-namespace Cli.Services
+namespace Cli.Services.Dotnet
 {
     internal class DotnetToolService : DotnetService
     {
